# src/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: Healthcare Management System API
  version: 1.0.0
  description: |
    HIPAA-compliant Healthcare Management System with comprehensive patient, appointment, and medical record management.
    
    ## Features
    - Patient management with full demographic information
    - Appointment scheduling and management
    - Medical record creation and retrieval with encryption
    - Patient consent management for HIPAA compliance
    - Secure authentication using JWT tokens
    - Comprehensive audit logging
    
    ## Security
    All endpoints require JWT Bearer token authentication.
    
    ## Compliance
    This API is designed to meet HIPAA compliance requirements including:
    - Data encryption at rest and in transit
    - Comprehensive audit trails
    - Access control and authorization
    - Patient consent management
  contact:
    name: Healthcare System Support
    email: support@healthcare.com
    url: https://healthcare.com
  license:
    name: Proprietary
    url: https://healthcare.com/license
servers:
  - url: http://localhost:8080
    description: Development Server
  - url: https://api.healthcare.com
    description: Production Server
security:
  - bearerAuth: []
...
paths:
  /api/v1/patients:
    post:
      tags:
        - Patient Management
      summary: Create a new patient
      description: Creates a new patient record in the system with all required information
      operationId: createPatient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientDTO'
      responses:
        '201':
          description: Patient created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientDTO'
        '400':
          description: Invalid input data
        '401':
          description: Unauthorized
        '409':
          description: Patient with email already exists
    get:
      tags:
        - Patient Management
      summary: Get all patients
      description: Retrieves a list of all patients in the system
      operationId: getAllPatients
      responses:
        '200':
          description: List of patients retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PatientDTO'
        '401':
          description: Unauthorized
  /api/v1/patients/{id}:
    get:
      tags:
        - Patient Management
      summary: Get patient by ID
      description: Retrieves a patient's complete information by their unique identifier
      operationId: getPatientById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Patient ID
      responses:
        '200':
          description: Patient found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientDTO'
        '401':
          description: Unauthorized
        '404':
          description: Patient not found
    put:
      tags:
        - Patient Management
      summary: Update patient information
      description: Updates an existing patient's information
      operationId: updatePatient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Patient ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientDTO'
      responses:
        '200':
          description: Patient updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientDTO'
        '400':
          description: Invalid input data
        '401':
          description: Unauthorized
        '404':
          description: Patient not found
    delete:
      tags:
        - Patient Management
      summary: Delete patient
      description: Soft deletes a patient by marking them as inactive
      operationId: deletePatient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Patient ID
      responses:
        '204':
          description: Patient deleted successfully
        '401':
          description: Unauthorized
        '404':
          description: Patient not found
...
  /api/v1/medical-records:
    post:
      tags:
        - Medical Record Management
      summary: Create a new medical record
      description: Creates a new medical record for a patient with diagnosis, treatment, and medication information
      operationId: createMedicalRecord
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MedicalRecordDTO'
      responses:
        '201':
          description: Medical record created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedicalRecordDTO'
        '400':
          description: Invalid input data
        '401':
          description: Unauthorized
        '404':
          description: Patient not found
  /api/v1/medical-records/{id}:
    get:
      tags:
        - Medical Record Management
      summary: Get medical record by ID
      description: Retrieves a specific medical record by its unique identifier
      operationId: getMedicalRecordById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Medical record ID
      responses:
        '200':
          description: Medical record found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedicalRecordDTO'
        '401':
          description: Unauthorized
        '403':
          description: Access denied - insufficient permissions
        '404':
          description: Medical record not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
  schemas:
    PatientDTO:
      type: object
      required:
        - firstName
        - lastName
        - dateOfBirth
        - gender
        - email
        - phoneNumber
      properties:
        id:
          type: integer
          format: int64
          description: Patient ID
          example: 1
        firstName:
          type: string
          maxLength: 100
          description: Patient's first name
          example: John
        lastName:
          type: string
          maxLength: 100
          description: Patient's last name
          example: Doe
        dateOfBirth:
          type: string
          format: date
          description: Patient's date of birth
          example: '1990-01-15'
        gender:
          type: string
          enum: [MALE, FEMALE, OTHER]
          description: Patient's gender
          example: MALE
        email:
          type: string
          format: email
          description: Patient's email address
          example: john.doe@email.com
        phoneNumber:
          type: string
          pattern: '^\+?[1-9]\d{1,14}$'
          description: Patient's phone number
          example: '+1234567890'
        bloodType:
          type: string
          maxLength: 50
          description: Patient's blood type
          example: 'A+'
        allergies:
          type: string
          description: Patient's known allergies
          example: 'Penicillin, Peanuts'
        medicalHistory:
          type: string
          description: Patient's medical history summary
        active:
          type: boolean
          description: Patient account active status
          example: true
...
    MedicalRecordDTO:
      type: object
      required:
        - patientId
        - recordType
        - recordDate
        - providerId
        - providerName
      properties:
        id:
          type: integer
          format: int64
          description: Medical record ID
          example: 1
        patientId:
          type: integer
          format: int64
          description: Patient ID
          example: 123
        recordType:
          type: string
          description: Type of medical record
          example: DIAGNOSIS
        recordDate:
          type: string
          format: date-time
          description: Date of the medical record
          example: '2024-01-15T10:30:00'
        providerId:
          type: string
          description: Healthcare provider identifier
          example: PROV123
        providerName:
          type: string
          description: Healthcare provider name
          example: Dr. Jane Smith
        diagnosis:
          type: string
          description: Diagnosis information
        treatment:
          type: string
          description: Treatment information
        medications:
          type: string
          description: Prescribed medications
        notes:
          type: string
          description: Additional notes
...